Disable progress spinner (for now)
To be replaced with something better.
