+++ typings.json
@@ -4,10 +4,7 @@
     "archy": "github:typings/typed-archy#3706c5e4df4efcc8cec8849a1c493cebd0830bea",
     "chalk": "github:typings/typed-chalk#a7e422c5455e70292e5675a727d43a7b05fc3e58",
     "columnify": "github:typings/typed-columnify#855f119ac0310bc0dc924b7c26dc1426e2aa43a5",
-    "elegant-spinner": "github:typings/typed-elegant-spinner#0779d7c419936b440e0951812c3fd9bf574bbc27",
-    "log-update": "github:typings/typed-log-update#e0e02e19ab2465c505337bdbb9f535a1d69fb817",
     "minimist": "github:typings/typed-minimist#9bfb8ebb1cf82abe2e16b00d6e0f59492224728e",
-    "promise-finally": "npm:promise-finally",
     "typings-core": "npm:typings-core",
     "update-notifier": "github:typings/typed-update-notifier#d3a027d90487f1b876157c75419f5c4a4a097a09",
     "wordwrap": "github:typings/typed-wordwrap#c5852a5a23954e721246001d4bc47d19c283cd33",
